requirements.txt
anyio==3.5.0
asgiref==3.5.0
click==8.0.3
fastapi==0.73.0
h11==0.13.0
idna==3.3
Pillow==9.0.1
pydantic==1.9.0
python-multipart==0.0.5
six==1.16.0
sniffio==1.2.0
starlette==0.17.1
typing_extensions==4.0.1
uvicorn==0.17.4
